+++ include/models/WaveNet.h
@@ -13,12 +13,12 @@ namespace MicroTorch
     class WaveNet : public BaseModel
     {
     public:
-        WaveNet(int input_size, int num_channels, int output_size, int kernel_size, std::vector<int> dilations, int stack_size, bool gated, Activation activation, float norm_mean, float norm_std) : 
+        WaveNet(size_t input_size, size_t num_channels, size_t output_size, size_t kernel_size, std::vector<size_t> dilations, size_t stack_size, bool gated, Activation activation, float norm_mean, float norm_std) : 
             BaseModel(norm_mean, norm_std), m_numChannels(num_channels), m_dilations(dilations), m_stackSize(stack_size), m_gated(gated), m_activation(activation),
             m_inputConv(input_size, num_channels, kernel_size, true, 1),
             m_outputLinear(num_channels, output_size, false)
         {
-            for(int k = 0; k < stack_size; k++)
+            for(size_t k = 0; k < stack_size; k++)
                 for(auto dilation: dilations)
                     m_blockStack.push_back( ResidualBlock(num_channels, kernel_size, dilation, true, true, gated, activation) );
         }
@@ -32,10 +32,10 @@ namespace MicroTorch
 
             RowMatrixXf skip_sum = Eigen::MatrixXf::Zero( m_numChannels, x.cols() );
             RowMatrixXf skip_y;
-            for(int k = 0; k < m_stackSize; k++)
-                for(int i = 0; i < m_dilations.size(); i++)
+            for(size_t k = 0; k < m_stackSize; k++)
+                for(size_t i = 0; i < m_dilations.size(); i++)
                 {
-                    int idx = k * m_dilations.size() + i;
+                    size_t idx = k * m_dilations.size() + i;
                     std::tie( y, skip_y ) = m_blockStack[idx].forward( y );
                     skip_sum.array() += skip_y.array();
                 }
@@ -47,10 +47,10 @@ namespace MicroTorch
         {
             auto conv_state_dict = state_dict[std::string("conv")].get<std::map<std::string, nlohmann::json>>();
             m_inputConv.loadStateDict( conv_state_dict );
-            for(int k = 0; k < m_stackSize; k++)
-                for(int i = 0; i < m_dilations.size(); i++)
+            for(size_t k = 0; k < m_stackSize; k++)
+                for(size_t i = 0; i < m_dilations.size(); i++)
                 {
-                    int idx = k * m_dilations.size() + i;
+                    size_t idx = k * m_dilations.size() + i;
                     auto block_state_dict = state_dict[std::string("blockStack.") + std::to_string(idx)].get<std::map<std::string, nlohmann::json>>();
                     m_blockStack[idx].loadStateDict( block_state_dict );
                 }
@@ -59,10 +59,10 @@ namespace MicroTorch
         }
 
     private:
-        int m_numChannels, m_stackSize;
+        size_t m_numChannels, m_stackSize;
         bool m_gated;
         Activation m_activation;
-        std::vector<int> m_dilations;
+        std::vector<size_t> m_dilations;
         CausalDilatedConv1d m_inputConv;
         std::vector<ResidualBlock> m_blockStack;
         Linear m_outputLinear;
